+++ test/cookbooks/test/attributes/default.rb
@@ -1,0 +1,2 @@
+# Which default MariaDB version to use for testing (can be overridden in kitchen.yml)
+default['mariadb_server_test_version'] = '10.3'